Removing assert failed message when the slave is not found when processing slaveDisconnected events
A slave can very often exit and be disconnected nearly at the same time,
which means the exit event coming from the launcher and the slaveDisconnected
event coming from attester can arrive in any order.

As a consequence, we should not log "ASSERT FAILED" if the slaveDisconnected
event is triggered after attester-launcher receives the exit event from the
launcher and has already removed the slave from its list.